Import spamd-devel.
This has the spamd files under "files" (here in pkgsrc-wip).
It is latest spamd from OpenBSD as of a few days ago.
And it has a few changes as seen in pkgsrc/mail/spamd.
This is incomplete.

Also it adds a few changes from me to choose SMTP messages (at build time).