Don't install spamlogd for now. (Need to finish that part.)
Improvements for spamd.8 manpage (need to get committed).

Banner without time. Need to make this optional.

Fix sync. There was corruption that I reported in April
and was patched and available in an OpenBSD errata.
(I update the version to the date of that errata patch.)

Add my workaround that doesn't do multicast. This needs to be fixed.
This means it will be broken for some using the multicast.
I mentioned it twice but I guess I need to file an official bug report.

Add feature for spamdb so it can select a -f filename to work on.
(I provided this patch, but I don't know if it was seen yet.)