Import my pkg_get script. The package is not done yet, but the script works. I have been slowly tweaking it and improving it for several months now on multiple pkgsrc systems.
It is a sh shell script.

It can download the pkg_summary database. And then uses that
to check what is installed, download packages, and install.

I put this in CVS to invite others to use it and improve it.
It is very short. It needs work. The script needs to be cleaned up.

Enjoy!