Development version of elinks (0.13) with a little code I wrote to allocate memory via libmm.  On NetBSD, this gives the speed of shared, anonymous memory and also contains nearly all of elinks' memory access within the shared memory pool, which is destroyed upon exit.  I like it much better like this. ;)
Also, I added some GNU code which is (or at least claims to be) a safe
wrapper for alloca() for additional safety.  Use of alloca() is
disabled in dcigettext.c; I did not have the energy to step through
it and replace all the instances of free().

If you use or are curious about using elinks, this is probably the way
to go.