Import marsyas-0.2.16 as wip/marsyas.
Initial check-in of work-in-progress (does not build to completion yet).

MARSYAS is a software framework that was created as part of my
research in Princeton University as a Phd graduate student. It is a
collection of classes written in C++ and JAVA for various sound
analysis and synthesis tasks. It started my first year in graduate school
(1998) when I rewrote various tools that I had been using in order to
make my life easier and also to code them the way I wanted them to be.

This library is made by me for my own research purposes. Anyone who
finds anything useful is welcome to use it, but I have no
responsibility whatsoever. I will be happy to answer any questions and
I will try to keep developing the library/toolkit. Until now
development is guided by what my research needs and whenever I find
some free time other stuff.

The basic goal for MARSYAS has been to design a flexible and easily
extensible system that allows rapid prototyping of audio
applications. MARSYAS is released as free software under the GNU
public licence hoping that people will actively contribute to its
development. Please send bugs, comments, additions, problems and
anything else to gtzan@@cs.cmu.edu. This version 0.2 is the third 
major rewrite of Marsyas and is still in experimental stage. Feedback 
is more than welcome. 

(This is the software behind MoodLogic and similar technologies/services)