Import mic-paren-3.7 as wip/mic-paren.
mic-paren.el is an extension and replacement to the packages paren.el
and stig-paren.el for Emacs. When mic-paren is active Emacs normal
parenthesis matching is deactivated. Instead parenthesis matching will
be performed as soon as the cursor is positioned at a parenthesis. The
matching parenthesis (or the entire expression between the
parentheses) is highlighted until the cursor is moved away from the
parenthesis.