Import tree-1.5.1.1nb2 as wip/tree.
Tree is a recursive directory listing program  that  produces  a  depth
indented  listing  of  files,  which  is colorized ala dircolors if the
LS_COLORS environment variable is set and output is to  tty.

Tree has various options for changing the information printed (size,
permissions, etc.) As well as plain text output, it is capable of
producing HTML output that includes links with an appropriate prefix.