Import hs-zlib-0.4.0.4 as wip/hs-zlib.
This package provides a pure interface for compressing and
decompressing streams of data represented as lazy ByteStrings. It uses
the zlib C library so it has high performance. It supports the "zlib",
"gzip" and "raw" compression formats.

It provides a convenient high level api suitable for most tasks and
for the few cases where more control is needed it provides access to
the full zlib feature set.