Import hslogger-1.0.5 as wip/hslogger.
hslogger is a logging framework for Haskell, roughly similar to
Python's logging module.

hslogger lets each log message have a priority and source be
associated with it. The programmer can then define global handlers
that route or filter messages based on the priority and
source. hslogger also has a syslog handler built in.