Import p5-aliased-0.22 as wip/p5-aliased.
aliased is simple in concept but is a rather
handy module. It loads the class you specify
and exports into your namespace a subroutine
that returns the class name. You can explicitly
alias the class to another name or, if you prefer
, you can do so implicitly. In the latter case,
the name of the subroutine is the last part of
the class name.