Import cacti-spine-0.8.7a as wip/cacti-spine.
Spine, formally Cactid, is a poller for Cacti that primarily strives to be
as fast as possible. For this reason it is written in native C, makes use of
POSIX threads, and is linked directly against the net-snmp library for minmumal
SNMP polling overhead. Spine is a replacement for the default cmd.php poller
so you must decide if using Spine makes sense for your installation.