Update to 2.7.2. Quoting the announcement messages:
2.7.0
-----
This is an update release of the filter.  The main purpose is to update
support for the signing practises test, an adjunct to DKIM for evaluating
whether or not a message should have been signed and, if it wasn't, what
to do with it.  In addition, a couple of features have been added.

The RELEASE_NOTES file fails to mention that all "ASP" references in the
configuration file were changed to "ADSP". This will be fixed in the
RELEASE_NOTES for the next version, but please update your configuration
files accordingly when installing this version.

The formal release notes entry:

2.7.0           2008/07/23
        Update to draft-ietf-dkim-ssp-04.  In doing so, rename "ASPDiscard"
                to "ADSPDiscard", "ASPNoSuchDomain" to "ADSPNoSuchDomain"
                and "SendASPReports" to "SendADSPReports" in the configuration
                file.
        Feature request #29738: Add "TrustSignaturesFrom" configuration
                file item allowing fine-grained control over third-party
                signature handling.
        Feature request #SF2018848: Add "LocalADSP" feature allowing
                policy assertions from domains known to have specific policies
                but which don't publish ADSP records.  Suggested by
                Bruno Kraychete da Costa.
        LIBDKIM: Fix an off-by-one overrun check in key and policy record
                decoding.  Problem noted by John Dickinson.

2.7.1
-----
This is a minor bug fix release.  Two signature processing issues have
been identified and resolved, and a rare issue having to do with crashes
inside OpenSSL has been addressed.

Thanks this time go out to Zbigniew Szalbot for his help tracking down the
OpenSSL problem and then testing the fix.

The formal release notes entry:

2.7.1           2008/08/27
        Set up required callbacks for OpenSSL thread-safety.  Problem
                noted by Zbigniew Szalbot.
        Disallow empty "t=" and "x=" tags.
        Return DKIM_STAT_KEYFAIL for various DNS key retrieval failures
                instead of DKIM_STAT_INTERNAL.

2.7.2
-----
This is a bugfix release, resolving a crash bug introduced in the previous
patch release.

The formal release notes entry:

2.7.2           2008/09/02
        Avoid memory leaks and infinite loops when releasing thread-specific
                memory.  Reported by Jeff Earickson.