Initial import of sunbird-0.9, updated version of time/sunbird: Doesn't work, just shows an empty Alert dialog on startup. Might need some additional files installed, but the installation framework is strange.
The Sunbird Project is a redesign of the Calendar component. The goal is to
produce a cross platform standalone calendar application based on Mozilla's
XUL user interface language. At the moment the Sunbird name is a project
name. It is not official and may change in the future.

The intended user is someone who uses Mozilla Firefox and Mozilla
Thunderbird and wants a calendar application based on Mozilla.